# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "blueeconomy-insurance",
  "version": "0.1.0"
 },
 "paths": {
...
  "/v1/quotes": {
   "post": {
    "tags": [
     "quotes"
    ],
    "summary": "Create Quote",
    "operationId": "create_quote_v1_quotes_post",
    "parameters": [
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QuoteCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Quote V1 Quotes Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/quotes:fromDeclaration": {
   "post": {
    "tags": [
     "quotes"
    ],
    "summary": "Quote From Declaration",
    "description": "Declaration-time attach (NTP VAS model): singlewindow posts the\nplatform declaration shape and receives a bound-ready quote offer.\nDeclaration dedupe: one quote per (declaration_ref, product) pair from a\ngiven source event; replays of the same source_event_id return the\noriginal quote.",
    "operationId": "quote_from_declaration_v1_quotes_fromDeclaration_post",
    "parameters": [
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FromDeclaration"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Quote From Declaration V1 Quotes Fromdeclaration Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/quotes/{quote_ref}": {
   "get": {
    "tags": [
     "quotes"
    ],
    "summary": "Get Quote",
    "operationId": "get_quote_v1_quotes__quote_ref__get",
    "parameters": [
     {
      "name": "quote_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Quote Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Quote V1 Quotes  Quote Ref  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/quotes/{quote_ref}:bind": {
   "post": {
    "tags": [
     "quotes"
    ],
    "summary": "Request Bind",
    "description": "Maker: request bind.",
    "operationId": "request_bind_v1_quotes__quote_ref__bind_post",
    "parameters": [
     {
      "name": "quote_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Quote Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Request Bind V1 Quotes  Quote Ref  Bind Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/quotes/{quote_ref}:bind-decision": {
   "post": {
    "tags": [
     "quotes"
    ],
    "summary": "Decide Bind",
    "description": "Checker: approve or decline. Checker != maker (service + DB CHECK).",
    "operationId": "decide_bind_v1_quotes__quote_ref__bind_decision_post",
    "parameters": [
     {
      "name": "quote_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Quote Ref"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BindDecisionIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Decide Bind V1 Quotes  Quote Ref  Bind Decision Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/quotes/{quote_ref}:issue": {
   "post": {
    "tags": [
     "quotes"
    ],
    "summary": "Issue Policy",
    "description": "Issue the policy as a signed W3C VC 2.0 credential and publish\ninsurance.policy.v1 (transactional outbox).",
    "operationId": "issue_policy_v1_quotes__quote_ref__issue_post",
    "parameters": [
     {
      "name": "quote_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Quote Ref"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IssueIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Issue Policy V1 Quotes  Quote Ref  Issue Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/quotes/{quote_ref}/policy": {
   "get": {
    "tags": [
     "quotes"
    ],
    "summary": "Get Policy For Quote",
    "operationId": "get_policy_for_quote_v1_quotes__quote_ref__policy_get",
    "parameters": [
     {
      "name": "quote_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Quote Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Policy For Quote V1 Quotes  Quote Ref  Policy Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BindDecisionIn": {
    "properties": {
     "decision": {
      "type": "string",
      "enum": [
       "BIND",
       "DECLINE"
      ],
      "title": "Decision"
     },
     "reason": {
      "type": "string",
      "maxLength": 2000,
      "title": "Reason",
      "default": ""
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "decision"
    ],
    "title": "BindDecisionIn"
   },
...
   "DeclarationLineIn": {
    "properties": {
     "hs_code": {
      "type": "string",
      "maxLength": 16,
      "title": "Hs Code",
      "default": ""
     },
     "description": {
      "type": "string",
      "maxLength": 2000,
      "title": "Description",
      "default": ""
     },
     "risk_class": {
      "type": "string",
      "maxLength": 64,
      "title": "Risk Class",
      "default": "general-cargo"
     },
     "customs_value_kobo": {
      "type": "integer",
      "maximum": 1000000000000000.0,
      "exclusiveMinimum": 0.0,
      "title": "Customs Value Kobo"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "customs_value_kobo"
    ],
    "title": "DeclarationLineIn",
    "description": "The platform declaration line shape consumed by the NTP VAS attach\nAPI (singlewindow declaration-time insurance offer)."
   },
...
   "FromDeclaration": {
    "properties": {
     "declaration_ref": {
      "type": "string",
      "maxLength": 64,
      "minLength": 1,
      "title": "Declaration Ref"
     },
     "source_event_id": {
      "type": "string",
      "maxLength": 128,
      "title": "Source Event Id",
      "default": ""
     },
     "product_code": {
      "type": "string",
      "maxLength": 64,
      "title": "Product Code",
      "default": "marine-cargo-single"
     },
     "corridor": {
      "type": "string",
      "maxLength": 128,
      "title": "Corridor",
      "default": ""
     },
     "consignee_name": {
      "type": "string",
      "maxLength": 256,
      "title": "Consignee Name",
      "default": ""
     },
     "consignee_tin": {
      "type": "string",
      "maxLength": 32,
      "title": "Consignee Tin",
      "default": ""
     },
     "occurred_at": {
      "type": "string",
      "format": "date-time",
      "title": "Occurred At"
     },
     "lines": {
      "items": {
       "$ref": "#/components/schemas/DeclarationLineIn"
      },
      "type": "array",
      "maxItems": 500,
      "minItems": 1,
      "title": "Lines"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "declaration_ref",
     "occurred_at",
     "lines"
    ],
    "title": "FromDeclaration"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "IssueIn": {
    "properties": {
     "inception_at": {
      "type": "string",
      "format": "date-time",
      "title": "Inception At"
     },
     "expiry_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expiry At"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "inception_at",
     "expiry_at"
    ],
    "title": "IssueIn"
   },
...
   "QuoteCreate": {
    "properties": {
     "product_code": {
      "type": "string",
      "pattern": "^[a-z0-9][a-z0-9._-]{0,63}$",
      "title": "Product Code"
     },
     "product_version": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Version"
     },
     "corridor": {
      "type": "string",
      "maxLength": 128,
      "title": "Corridor",
      "default": ""
     },
     "declaration_ref": {
      "type": "string",
      "maxLength": 64,
      "title": "Declaration Ref",
      "default": ""
     },
     "assured_name": {
      "type": "string",
      "maxLength": 256,
      "title": "Assured Name",
      "default": ""
     },
     "assured_tin": {
      "type": "string",
      "maxLength": 32,
      "title": "Assured Tin",
      "default": ""
     },
     "lines": {
      "items": {
       "$ref": "#/components/schemas/QuoteLineIn"
      },
      "type": "array",
      "maxItems": 500,
      "minItems": 1,
      "title": "Lines"
     },
     "expected_premium_kobo": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Expected Premium Kobo"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "product_code",
     "lines"
    ],
    "title": "QuoteCreate"
   },
   "QuoteLineIn": {
    "properties": {
     "description": {
      "type": "string",
      "maxLength": 2000,
      "title": "Description",
      "default": ""
     },
     "hs_code": {
      "type": "string",
      "maxLength": 16,
      "title": "Hs Code",
      "default": ""
     },
     "risk_class": {
      "type": "string",
      "maxLength": 64,
      "minLength": 1,
      "title": "Risk Class"
     },
     "insured_value_kobo": {
      "type": "integer",
      "maximum": 1000000000000000.0,
      "exclusiveMinimum": 0.0,
      "title": "Insured Value Kobo"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "risk_class",
     "insured_value_kobo"
    ],
    "title": "QuoteLineIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}